Improve: Refactor code  - Add handle error for admin pages  - Remove current_user from user list
