+++ app/controllers/concerns/api/response.rb
@@ -1,0 +1,11 @@
+module Api
+  module Response
+    extend ActiveSupport::Concern
+
+    included do
+      def json_response(object, status = :ok)
+        render json: object, status: status, root: false
+      end
+    end
+  end
+end
