--- app/controllers/concerns/response.rb
@@ -1,9 +1,0 @@
-module Response
-  extend ActiveSupport::Concern
-
-  included do
-    def json_response(object, status = :ok)
-      render json: object, status: status, root: false
-    end
-  end
-end
