+++ app/controllers/users_controller.rb
@@ -2,7 +2,7 @@ class UsersController < ApplicationController
   before_action :set_user, only: [:show, :edit, :update, :destroy]
 
   def index
-    @users = User.all
+    @users = User.where.not(id: current_user.id).all
   end
 
   def show
@@ -41,7 +41,7 @@ class UsersController < ApplicationController
   private
 
   def set_user
-    @user = User.find(params[:id])
+    @user = User.where.not(id: current_user.id).find(params[:id])
   end
 
   def user_params
